# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Graduation Project Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profiles/me:
    get:
      operationId: getMyProfile
      tags: [profiles]
      summary: Get my student profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
    put:
      operationId: updateMyProfile
      tags: [profiles]
      summary: Update my student profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
  /profiles/{userId}:
    get:
      operationId: getProfile
      tags: [profiles]
      summary: Get a student profile by user ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
  /profiles:
    get:
      operationId: listProfiles
      tags: [profiles]
      summary: List student profiles with search
      parameters:
        - name: skills
          in: query
          schema:
            type: string
        - name: minGpa
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Profiles list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentProfile"
...
  /supervisor-requests:
    get:
      operationId: listSupervisorRequests
      tags: [supervisors]
      summary: List supervisor requests
      responses:
        "200":
          description: Requests list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupervisorRequest"
    post:
      operationId: sendSupervisorRequest
      tags: [supervisors]
      summary: Send supervision request to a supervisor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendSupervisorRequestBody"
      responses:
        "201":
          description: Request sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupervisorRequest"
  /supervisor-requests/{id}/accept:
    post:
      operationId: acceptSupervisorRequest
      tags: [supervisors]
      summary: Accept supervision request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /supervisor-requests/{id}/reject:
    post:
      operationId: rejectSupervisorRequest
      tags: [supervisors]
      summary: Reject supervision request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /coordinator/assign-supervisor:
    post:
      operationId: coordinatorAssignSupervisor
      tags: [supervisors]
      summary: Coordinator manually assigns supervisor to team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignSupervisorBody"
      responses:
        "200":
          description: Assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /submissions:
    get:
      operationId: listSubmissions
      tags: [submissions]
      summary: List submissions
      responses:
        "200":
          description: Submissions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
    post:
      operationId: createSubmission
      tags: [submissions]
      summary: Submit a deliverable for a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubmissionBody"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /submissions/{id}/review:
    post:
      operationId: reviewSubmission
      tags: [submissions]
      summary: Review a submission (supervisor)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewSubmissionBody"
      responses:
        "200":
          description: Reviewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, supervisor, coordinator]
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, createdAt]
...
    StudentProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        gpa:
          type: number
          nullable: true
        skills:
          type: string
          nullable: true
        interests:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        user:
          $ref: "#/components/schemas/User"
      required: [id, userId, user]
    UpdateProfileBody:
      type: object
      properties:
        gpa:
          type: number
          nullable: true
        skills:
          type: string
          nullable: true
        interests:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
    Team:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        projectTitle:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [forming, active, supervised, completed]
        leaderId:
          type: integer
        supervisorId:
          type: integer
          nullable: true
        currentPhase:
          type: string
          enum: [proposal, progress, final]
          nullable: true
        createdAt:
          type: string
          format: date-time
        leader:
          $ref: "#/components/schemas/User"
        supervisor:
          $ref: "#/components/schemas/User"
          nullable: true
        memberCount:
          type: integer
      required: [id, name, status, leaderId, createdAt, leader, memberCount]
...
    SupervisorRequest:
      type: object
      properties:
        id:
          type: integer
        teamId:
          type: integer
        supervisorId:
          type: integer
        status:
          type: string
          enum: [pending, accepted, rejected]
        message:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        team:
          $ref: "#/components/schemas/Team"
        supervisor:
          $ref: "#/components/schemas/User"
      required: [id, teamId, supervisorId, status, createdAt, team, supervisor]
    SendSupervisorRequestBody:
      type: object
      properties:
        supervisorId:
          type: integer
        message:
          type: string
          nullable: true
      required: [supervisorId]
    AssignSupervisorBody:
      type: object
      properties:
        teamId:
          type: integer
        supervisorId:
          type: integer
      required: [teamId, supervisorId]
...
    Task:
      type: object
      properties:
        id:
          type: integer
        teamId:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        deadline:
          type: string
          format: date-time
          nullable: true
        phase:
          type: string
          enum: [proposal, progress, final]
        status:
          type: string
          enum: [pending, submitted, reviewed]
        createdAt:
          type: string
          format: date-time
        team:
          $ref: "#/components/schemas/Team"
      required: [id, teamId, title, phase, status, createdAt, team]
...
    Submission:
      type: object
      properties:
        id:
          type: integer
        taskId:
          type: integer
        submittedById:
          type: integer
        fileUrl:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, approved, rejected]
        feedback:
          type: string
          nullable: true
        submittedAt:
          type: string
          format: date-time
        task:
          $ref: "#/components/schemas/Task"
        submittedBy:
          $ref: "#/components/schemas/User"
      required: [id, taskId, submittedById, status, submittedAt, task, submittedBy]
    CreateSubmissionBody:
      type: object
      properties:
        taskId:
          type: integer
        fileUrl:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [taskId]
    ReviewSubmissionBody:
      type: object
      properties:
        status:
          type: string
          enum: [approved, rejected]
        feedback:
          type: string
          nullable: true
      required: [status]
...