# lib/api-zod/src/generated/types/assignSupervisorBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Graduation Project Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface AssignSupervisorBody {
  teamId: number;
  supervisorId: number;
}